+++ changelog.md
@@ -3,6 +3,7 @@
 
 
  - [feature] added Fränkisch to list of valid languages
+ - [feature] fair ranking algorithm
  - [feature] add allowed languages
  - [feature] no more pr ordering in backend / two order fields for assigned and unassigned prs
  - [feature] add IT lang config